+++ generator-prisma-client/nextjs-starter-turbopack/.eslintrc.json
