+++ generator-prisma-client/nextjs-starter-turbopack/app/api/route.ts
