+++ generator-prisma-client/nextjs-starter-turbopack/app/favicon.ico
