+++ generator-prisma-client/nextjs-starter-turbopack/app/globals.css
