+++ generator-prisma-client/nextjs-starter-turbopack/app/page.tsx
@@ -7,12 +7,12 @@ export default async function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-start p-8 md:p-24 bg-white dark:bg-slate-900">
       <div className="text-center mb-12">
-        <h1 className="text-4xl font-bold text-slate-800 dark:text-white mb-4">
-          Quotes from Prisma Postgres
+      <h1 className="text-2xl font-bold text-slate-800 dark:text-white mb-4">
+          Using: Next.js 15, Turbopack, Node.js, <pre className="inline-block">prisma-client</pre>
         </h1>
 
         <p className="text-xl text-slate-600 dark:text-slate-300">
-          A collection of inspiring opinions and fascinating facts
+          Quotes: A collection of inspiring opinions and fascinating facts
         </p>
       </div>
       <div className="text-center mb-12">
