+++ generator-prisma-client/nextjs-starter-turbopack/components/quotes.tsx
