+++ generator-prisma-client/nextjs-starter-turbopack/lib/prisma-enums.ts
@@ -1,0 +1,1 @@
+export * from './generated/prisma/enums'
