--- prisma-postgres/nextjs-starter-turbopack/lib/prisma-enums.ts
@@ -1,1 +1,0 @@
-export * from './.generated/prisma/enums'
